Get rid of Ada-enabled sys-devel/gcc, Ada packages, and also CKermit.
GNAT has better maintenance now in the ‘gentoo’ repo. The binary
community distribution can be installed in opt in a reasonably usable
way. A usable amount of GNAT software can be built from sources, if
one prefers.

CKermit fiddles with compiler internals and is broken by changes in
gcc. It is basically an MSDOS-style program that isn’t properly
written for Posix.